Observations to grid designation to create a data cube

The function designates observations to cells of a given grid to create an aggregated data cube.

grid_designation(
  observations,
  grid,
  id_col = "row_names",
  seed = NA,
  aggregate = TRUE,
  randomisation = c("uniform", "normal"),
  p_norm = ifelse(tolower(randomisation[1]) == "uniform", NA, 0.95)
)

Arguments

observations

An sf object with POINT geometry and a coordinateUncertaintyInMeters column. If this column is not present, the function will assume no (zero meters) uncertainty around the observation points.

grid

An sf object with POLYGON geometry (usually a grid) to which observations should be designated.

id_col

The column name of the column with unique ids for each grid cell. If "row_names" (the default), a new column id is created were the row names represent the unique ids.

seed

The seed for random number generation to make results reproducible. If NA (the default), no seed is used.

aggregate

Logical. If TRUE (default), return data cube in aggregated form (grid with number of observations per grid cell). Otherwise return sampled points in uncertainty circle.

randomisation

"uniform" or "normal". Randomisation method used for sampling within uncertainty circle around each observation. By default "uniform" which means each point uncertainty circle has an equal probability to be selected. The other option is "normal" where a point is sampled from a bivariate Normal distribution with means equal to the observation point and the variance equal to (-coordinateUncertaintyInMeters^2) / (2 * log(1 - p_norm)) such that p_norm % of all possible samples from this Normal distribution fall within the uncertainty circle.

p_norm

A numeric value between 0 and 1. Only used if randomisation = "normal". The proportion of all possible samples from a a bivariate Normal distribution that fall within the uncertainty circle. If normal randomisation is used and no value is given, the default p_norm value is 0.95.

Value

In case of aggregate = TRUE, an sf object with POLYGON geometry containing the locations of the grid cells, an n column with the number of observations per grid cell, and a min_coord_uncertainty column containing the minimal coordinate uncertainty per grid cell. In case of aggregate = FALSE, an sf object with POINT geometry containing the locations of the sampled observations within the uncertainty circle, and a coordinateUncertaintyInMeters column containing the coordinate uncertainty for each observation.

Examples

library(simcuber)
library(sf)
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is TRUE
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

set.seed(123)

# Create four random points
n_points <- 4
xlim <- c(3841000, 3842000)
ylim <- c(3110000, 3112000)
coordinate_uncertainty <- rgamma(n_points, shape = 5, rate = 0.1)

observations_sf <- data.frame(
  lat = runif(n_points, ylim[1], ylim[2]),
  long = runif(n_points, xlim[1], xlim[2]),
  coordinateUncertaintyInMeters = coordinate_uncertainty
) %>%
  st_as_sf(coords = c("long", "lat"), crs = 3035)

# Add buffer uncertainty in meters around points
observations_buffered <- observations_sf %>%
  st_buffer(observations_sf$coordinateUncertaintyInMeters)

# Create grid
grid_df <- st_make_grid(
  observations_buffered,
  square = TRUE,
  cellsize = c(200, 200)
) %>%
  st_sf()

# Create occurrence cube
grid_designation(
  observations = observations_sf,
  grid = grid_df,
  seed = 123
)
#> Simple feature collection with 30 features and 3 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 3840994 ymin: 3110833 xmax: 3841994 ymax: 3112033
#> Projected CRS: ETRS89-extended / LAEA Europe
#> # A tibble: 30 × 4
#>    id        n min_coord_uncertainty                                    geometry
#>  * <chr> <int>                 <dbl>                               <POLYGON [m]>
#>  1 12        1                  16.3 ((3841194 3111233, 3841394 3111233, 384139…
#>  2 26        1                  33.9 ((3840994 3111833, 3841194 3111833, 384119…
#>  3 5         1                  73.8 ((3841794 3110833, 3841994 3110833, 384199…
#>  4 6         1                  47.8 ((3840994 3111033, 3841194 3111033, 384119…
#>  5 1         0                  NA   ((3840994 3110833, 3841194 3110833, 384119…
#>  6 2         0                  NA   ((3841194 3110833, 3841394 3110833, 384139…
#>  7 3         0                  NA   ((3841394 3110833, 3841594 3110833, 384159…
#>  8 4         0                  NA   ((3841594 3110833, 3841794 3110833, 384179…
#>  9 7         0                  NA   ((3841194 3111033, 3841394 3111033, 384139…
#> 10 8         0                  NA   ((3841394 3111033, 3841594 3111033, 384159…
#> # ℹ 20 more rows