Sample from a circle using the Uniform distribution

The function samples occurrences of a species within the uncertainty circle around each observation assuming a Uniform distribution.

sample_from_uniform_circle(observations, seed = NA)

Arguments

observations

An sf object with POINT geometry and a coordinateUncertaintyInMeters column. If this column is not present, the function will assume no (zero meters) uncertainty around the observation points.

seed

A positive numeric value. The seed for random number generation to make results reproducible. If NA (the default), no seed is used.

Value

An sf object with POINT geometry containing the locations of the sampled occurrences and a coordinateUncertaintyInMeters column containing the coordinate uncertainty for each observation.

Examples

library(simcuber)
library(sf)

set.seed(123)

# Create four random points
n_points <- 4
xlim <- c(3841000, 3842000)
ylim <- c(3110000, 3112000)
coordinate_uncertainty <- rgamma(n_points, shape = 5, rate = 0.1)

observations_sf <- data.frame(
  lat = runif(n_points, ylim[1], ylim[2]),
  long = runif(n_points, xlim[1], xlim[2]),
  coordinateUncertaintyInMeters = coordinate_uncertainty
) %>%
  st_as_sf(coords = c("long", "lat"), crs = 3035)

# Sample points within uncertainty circles according to uniform rules
sample_from_uniform_circle(
  observations = observations_sf,
  seed = 123
)
#> Simple feature collection with 4 features and 1 field
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 3841076 ymin: 3110891 xmax: 3841904 ymax: 3111946
#> Projected CRS: ETRS89-extended / LAEA Europe
#>   coordinateUncertaintyInMeters                geometry
#> 1                      33.89585 POINT (3841095 3111946)
#> 2                      73.78957 POINT (3841904 3110891)
#> 3                      16.29561 POINT (3841236 3111362)
#> 4                      47.78440 POINT (3841076 3111115)